requirements.txt
ctransformers==0.2.5
sentence_transformers
pinecone-client
langchain
flask
pypdf
python-dotenv
-e .